Astrophobia | Dumpster Fire (Pre-Alpha) v0.0.3
Dumpster Fire v0.0.3 is here!! (this is barely even a dumpster, let alone a dumpster fire)

Some key changes in this pseudo-release (not planning on releasing pre-alphas) are the introduction of scrolling text and organized spacing, attributes found in other typical command-line RPGs. Some smaller changes are the temporary removal of the commented *bones* of character creation, until I feel not completely drained and do not have hours of worksheets of math to do, that is. Thanks for keeping caught up! New release with ACTUAL REAL GAMEPLAY!! should be available by next weekend at the very latest. Buh-bye!